pyproject.toml
[project]
name = "gkendpoints"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-auth>=2.38.0",
    "google-cloud-compute>=1.29.0",
    "google-cloud-container>=2.56.0",
    "kubernetes>=32.0.1",
]
